[Feat] - not implementat

+++ Shared/Features/Product/ProductEntity.cs
@@ -42,7 +42,7 @@ namespace Shared.Features
         public int MaxCount { get; set; }
 
         [Column("info_count")]
-        public long InfoCount { get; set; }
+        public int InfoCount { get; set; }
 
         [Column("price")]
         public decimal Price { get; set; }
